Added better setUp to RTL433 error tests. The test class now checks for PID file and checks PIDs.
Moved the sleep funcion in the main loop. This doesn't really work
as the programme waits 15s, before rtl_433 has data but data is
received the outer loop is not entered.

The data-gathering block enters the 'while not empty' block it
stays there. This should probably change to 'if not
stdout_queue.empty()'.

Added and commented out a lot of diagnostic print statements.

Added subprocess PID file creation.

Added skeleton tests for PID checking.

TODO
* Add PID tests for rtl_433_2sqlite and rtl_433 (subprocess)
* Close threads gracefully.
* Replace 'while not stdout_queue.empty()' with
  'if not stdout_queue.empty()'
* Solve infiinte loop - stuck in 'a' [done]
* Get tests to run gracefully with PID checking/deletion [done]
* Add subprocess PID to PID file. - [done]