Fixing tests.
testBlankresponse checks that get is called the
correct number of times but no database calls.

testGoodBadGood checks that get is called the correct number of
times, the database is called one less (due to bad input) and
is *meant* to check the database calls were correct. ***
Does not work as it becomes JSON inbetween filereader and
database call.

TODO
* Get tests to pass before merging PID
* Add PID tests for rtl_433_2sqlite and rtl_433 (subprocess)
* Close threads gracefully.
* Replace 'while not stdout_queue.empty()' with
  'if not stdout_queue.empty()'